refactor pixel_mat
